fix: improve determineIfShotWillHit method
