+++ src/main/kotlin/com/dp/advancedgunnerycontrol/weaponais/AGCUitls.kt
@@ -4,9 +4,9 @@ package com.dp.advancedgunnerycontrol.weaponais
 
 import com.dp.advancedgunnerycontrol.WeaponControlPlugin
 import com.dp.advancedgunnerycontrol.settings.Settings
-import com.dp.advancedgunnerycontrol.utils.effectiveCollRadius
 import com.fs.starfarer.api.Global
 import com.fs.starfarer.api.combat.*
+import org.lazywizard.lazylib.MathUtils
 import org.lazywizard.lazylib.ext.minus
 import org.lazywizard.lazylib.ext.plus
 import org.lwjgl.util.vector.Vector2f
@@ -243,4 +243,36 @@ fun mapBooleanToSpecificString(boolValue: Boolean, trueString: String, falseStri
 // Why doesn't Vector2f support this naturally? Note: infix and _ rather than operator in case this ever gets added
 internal infix fun Vector2f.times_(d: Float): Vector2f {
     return Vector2f(d * x, d * y)
+}
+
+
+/**
+ * @return approximate angular distance of target from current weapon facing in rad
+ * note: approximation works well for small values and is off by a factor of PI/2 for 180°
+ * @param entity: Relative coordinates (velocity-compensated)
+ */
+fun angularDistanceFromWeapon(entity: Vector2f, weapon: WeaponAPI): Float {
+    val weaponDirection = vectorFromAngleDeg(weapon.currAngle)
+    val distance = entity - weapon.location
+    val entityDirection = distance times_ (1f / distance.length())
+    return (weaponDirection - entityDirection).length()
+}
+
+fun linearDistanceFromWeapon(entity: Vector2f, weapon: WeaponAPI): Float {
+    return (weapon.location - entity).length()
+}
+/**
+ * @param entity: In relative coordinates
+ * @param collRadius: Include any tolerances in here
+ * @param aimPoint: Point the weapon is aiming at, deduced from current weapon facing if not provided
+ */
+fun determineIfShotWillHit(entity: Vector2f, collRadius: Float, weapon: WeaponAPI, aimPoint: Vector2f? = null) : Boolean{
+    val p = aimPoint?.minus(weapon.location) ?: vectorFromAngleDeg(weapon.currAngle)
+    val e = entity - weapon.location
+    val x = (e.x * p.x + e.y * p.y) / (p.x * p.x + p.y * p.y)
+    return x > 0 && MathUtils.getDistanceSquared(p.times_(x), e) < collRadius * collRadius
+}
+
+fun effectiveCollRadius(entity: CombatEntityAPI) : Float{
+    return entity.collisionRadius * Settings.collisionRadiusMultiplier()
 }
